APIGateway/requirements.txt
Flask
requests
pika
pymssql
prometheus_client
seqlog
grpcio
grpcio-tools